// src/app/models/profile-general-data.model.ts
export class profileGeneralData{
    purchGroup : any;
    plantCode : any;
    title : any;
    vendorName1 : any;
    vendorName2 : any;
    streetLine1 : any;
    streetLine2 : any;
    streetLine3 : any;
    streetLine4 : any;
    postalCode  : any;
    countryKey : any;
    region: any;
    city : any;
    telephoneNo : any;
    mobileNo: any;
    faxNo : any;
    vendorEmail: any;
    bankName : any;
    branch: any;
    bankAccountNo : any;
    swiftCode: any;
    ifscNo : any;
    pfRegisterationNo: any;
    ESICRegisterationNo: any;
    VATRegisterationNo: any;
    panNo: any;
    serviceTaxRegisterationNo: any;
    CSTNo: any;
    LSTNo: any;
    GSTRegisterationNo: any;
    taxClassification: any;
    TDSExcemptionNo: any;
    excemptionPercentage: any;
    // TDSExcemptionNo2: any;
    // excemptionPercentage2: any;
    excemptFrom: any;
    excemptTo: any;
    MSMEStatus: any;
    MSMENo: any;
    fileUpload1: any;
    businessType: any;
    industryType: any;
    promoters: any;
    top5Clients: any;
    turnOver: any;
    clientRef: any;
    complianceCategory: any;
    constitutionBus: any;
    fileUpload2: any;
    firstName: any;
    lastName: any;
    emailID: any;
    phoneNo: any;
    DOB: any;
    gender: any;
    address: any;
    city1: any;
    state: any;
    country: any;
    Hobbies: any;
    companyName: any;
    parentCompany: any;
    globalHeadOffice: any;
    IndiaHeadOffice: any;
    UploadPortfolio: any;
    website: any;
    manufacturingSite: any;
    mfgCapacity: any;
    yearofEshtablihsment: any;
    companyDescription: any;
    state1: any;
    country1: any;
    revenueLY: any;
    PATLY: any;
    LYSpend: any;
    EmployeeSize: any;
    NameL1: any;
    emailIDL1: any;
    contactNoL1: any;
    nameL2: any;
    emailIDL2: any;
    contactNoL2: any;
    BCCLAssociation: any;
    BCCLannualSpend: any;
    groupCO: any;
    marketUpdate: any;
}